put calculation function in timeCounter module
